feat(modularization): S-7 -- extract scheduled tasks frontend to web/modules/schedules.js
Moves the full scheduled-tasks UI section (~929 lines) from the monolithic
web/app.js into web/modules/schedules.js as an ES module.

- Exports: initSchedules, loadSchedules, loadScheduleAgents, openEditSchedule,
  getScheduleCron (consumed by app.js)
- DI pattern: openModal/closeModal injected via initSchedules({...})
- Inline local helpers: pauseIcon/playIcon/trashIcon (SVGs were referenced
  from app.js line scope; inlined to avoid ReferenceError at runtime)
- Imports: showToast (toast.js), t (i18n.js), avatarBust (agents.js)
- Updated schedule-run-now.test.ts to read from schedules.js
- Verified: node --check, npm run build, 2873 vitest tests passing,
  Playwright headless smoke 0 JS errors (tasks/agents/kanban pages all load)
